Time: 6 ms (22.41%) | Memory: 45.3 MB (54.22%) - LeetSync

+++ 137-single-number-ii/single-number-ii.java
@@ -18,7 +18,7 @@ class Solution {
         {
             if(map.get(key)==1)
             {
-                return (int)key;
+                return key;
             }
         }
         return 0;
